requirements.txt
fastapi
minio
uvicorn
celery[redis]
pydantic_settings
ffmpeg-python
python-multipart
